backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.25
geoalchemy2==0.14.3
alembic==1.13.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-dotenv==1.0.1
celery==5.3.6
redis==5.0.1
anthropic==0.18.1
gdal==3.6.4
rasterio==1.3.9
fiona==1.9.5
shapely==2.0.3
pyproj==3.6.1
numpy==1.26.4
scipy==1.12.0
opencv-python-headless==4.9.0.80
scikit-image==0.22.0
pytesseract==0.3.10
reportlab==4.1.0
weasyprint==62.3
folium==0.16.0
httpx==0.26.0
pydantic==2.6.1
pydantic-settings==2.1.0
difflib2==1.0.0
pytest==8.0.1
pytest-asyncio==0.23.5
